Move `ReplaceFile` function to `util` package from `driver` package
